+++ packages/moonstone/IncrementSlider/IncrementSliderButton.js
@@ -1,7 +1,10 @@
 import kind from '@enact/core/kind';
+import Holdable from '@enact/ui/Holdable';
 import IconButton from '../IconButton';
 import onlyUpdateForKeys from 'recompose/onlyUpdateForKeys';
 import React from 'react';
+
+const HoldableIconButton = Holdable(IconButton);
 
 /**
  * {@link moonstone/IncrementSlider.IncrementSliderButton} is an
@@ -26,9 +29,18 @@ const IncrementSliderButtonBase = kind({
 		}
 	},
 
-	render: (props) => (
-		<IconButton {...props} backgroundOpacity="transparent" small />
-	)
+	render: ({onClick, ...rest}) => {
+		return (
+			<HoldableIconButton
+				{...rest}
+				backgroundOpacity="transparent"
+				onClick={onClick}
+				onHold={onClick}
+				onHoldPulse={onClick}
+				small
+			/>
+		);
+	}
 });
 
 const OnlyUpdate = onlyUpdateForKeys(['children', 'disabled', 'spotlightDisabled']);